+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Item View position can now be swapped between the left and right panels.
 - EIV's settings can now be modified in Mod Menu.
 - Adjusted "Entity Fighting" View Type to take up less of the screen.
+- Added a toggle for EIV to the client settings screen.